refactor missing entity class to ffe_entity
